+++ routes/users.js
@@ -2,7 +2,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 
-const User = require('../models/user');
+const User = require('../models/User');
 const { ValidationError } = require('../errors');
 
 const router = express.Router();
